--- app/src/main/java/com/example/commerceapp/domain/usecases/user/GetUserInfoUseCase.kt
@@ -1,6 +1,0 @@
-package com.example.commerceapp.domain.usecases.user
-
-import com.example.commerceapp.domain.repository.UserInfoRepository
-
-class GetUserInfoUseCase(userInfoRepository: UserInfoRepository) {
-}
