+++ app/src/main/java/com/example/commerceapp/domain/usecases/user/GetUserPreviewInfoUseCase.kt
@@ -1,19 +1,19 @@
-package com.example.commerceapp.domain.usecases.user.withlist
+package com.example.commerceapp.domain.usecases.user
 
 import com.example.commerceapp.domain.extension.mapToResultEntity
 import com.example.commerceapp.domain.model.common.ErrorHandler
 import com.example.commerceapp.domain.model.common.RequestParam
 import com.example.commerceapp.domain.model.common.ResultEntity
-import com.example.commerceapp.domain.model.product.Product
+import com.example.commerceapp.domain.model.user.UserPreview
 import com.example.commerceapp.domain.repository.UserRepository
 import com.example.commerceapp.domain.usecases.base.BaseFlowUseCase
 import kotlinx.coroutines.flow.Flow
 
-data class GetWishListUseCase(
+class GetUserPreviewInfoUseCase(
     val repository: UserRepository,
     private val loginExceptionHandler: ErrorHandler
-) : BaseFlowUseCase<List<Product>>() {
-    override suspend fun invoke(parameters: RequestParam): Flow<ResultEntity<List<Product>>> =
-        repository.getWishList(parameters).mapToResultEntity(loginExceptionHandler)
+) : BaseFlowUseCase<UserPreview>() {
+    override suspend fun invoke(parameters: RequestParam): Flow<ResultEntity<UserPreview>> =
+        repository.getMyPreviewInfo(parameters).mapToResultEntity(loginExceptionHandler)
 
 }
